Now that we're not using the entities pattern, move the file to a better spot.
